Conflicts:
	core/baseLanguage/baseLanguage/languageModels/behavior.mps
	core/baseLanguage/baseLanguage/languageModels/editor.mps
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/behavior/.generated
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/behavior/.hash.-9hqq74u2egvra0pssg6x0tg7oezcawa
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/behavior/.hash.2hiot7rv2fgr467cwoob3lxuj6ipd8b
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/behavior/.hash.alrrwlgtux43afg763vbfjxwr98xpdh
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/editor/.generated
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/editor/.hash.5o6f0oojfmfahcpx6aw4rqpsi3nkjhq
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/editor/.hash.63j6u86emewcfd4n1sqt5jrpwpq1ffp
	core/baseLanguage/baseLanguage/source_gen.caches/jetbrains/mps/baseLanguage/editor/.hash.6d9eci4pu9tp1bjvqauo45pl47cascl
	core/languageDesign/plugin/generator/baseLanguage/template/main@generator.mps
	core/languageDesign/plugin/source_gen.caches/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/.generated
	core/languageDesign/plugin/source_gen.caches/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/.hash.-46bk7hipwpuntiw75olthtqnexz5xkj
	core/languageDesign/plugin/source_gen.caches/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/.hash.1f2jn5925mse76uwxz6anxlppgscefg
	core/languageDesign/plugin/source_gen.caches/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/.hash.dp7gxcmihz85r982rlggxkcj8nj4p3c
	core/languageDesign/plugin/source_gen/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/QueriesGenerated.java
	core/languageDesign/plugin/source_gen/jetbrains/mps/lang/plugin/generator/baseLanguage/template/main/trace.info

+++ core/kernel/source_gen/jetbrains/mps/util/LongestCommonSubsequenceFinder.java
@@ -118,11 +118,11 @@ __switch__:
           };
         }
       }.invoke()))) {
-        if (!(i == f._0() && j == f._1())) {
-          ListSequence.fromList(myDifferentIndices).addElement(MultiTuple.<Tuples._2<Integer, Integer>,Tuples._2<Integer, Integer>>from(MultiTuple.<Integer,Integer>from(i, f._0()), MultiTuple.<Integer,Integer>from(j, f._1())));
+        if (!(i == (int) f._0() && j == (int) f._1())) {
+          ListSequence.fromList(myDifferentIndices).addElement(MultiTuple.<Tuples._2<Integer, Integer>,Tuples._2<Integer, Integer>>from(MultiTuple.<Integer,Integer>from(i, (int) f._0()), MultiTuple.<Integer,Integer>from(j, (int) f._1())));
         }
-        i = f._0() + 1;
-        j = f._1() + 1;
+        i = (int) f._0() + 1;
+        j = (int) f._1() + 1;
       }
     }
     return myDifferentIndices;
@@ -135,7 +135,7 @@ __switch__:
     if (myDifferentSubsequences == null) {
       myDifferentSubsequences = ListSequence.fromList(getDifferentIndices()).<Tuples._2<Iterable<E>, Iterable<E>>>select(new ISelector<Tuples._2<Tuples._2<Integer, Integer>, Tuples._2<Integer, Integer>>, Tuples._2<Iterable<E>, Iterable<E>>>() {
         public Tuples._2<Iterable<E>, Iterable<E>> select(Tuples._2<Tuples._2<Integer, Integer>, Tuples._2<Integer, Integer>> pair) {
-          return MultiTuple.<Iterable<E>,Iterable<E>>from(ListSequence.fromList(myA).page(pair._0()._0(), pair._0()._1()), ListSequence.fromList(myB).page(pair._1()._0(), pair._1()._1()));
+          return MultiTuple.<Iterable<E>,Iterable<E>>from(ListSequence.fromList(myA).page((int) pair._0()._0(), (int) pair._0()._1()), ListSequence.fromList(myB).page((int) pair._1()._0(), (int) pair._1()._1()));
         }
       }).toListSequence();
     }
